% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_excel.R
\name{create_gap_workbook}
\alias{create_gap_workbook}
\alias{as_row}
\title{Functions for producing Excel tables in Global style}
\usage{
create_gap_workbook(...)

as_row(x)
}
\arguments{
\item{...}{Other arguments passed down to \code{createWorkbook}.}

\item{x}{Vector to coerce into a \code{data.frame} object.}
}
\value{
Workbook object.

\itemize{
\item \code{as_row} returns a \code{data.frame} object that occupies one row.
}
}
\description{
The function \code{create_gap_workbook} is a wrapper for \code{op}enxlsx::createWorkbook\verb{that sets workbook font to Franklin Gothic Book, 9 point font. The function}as_row\verb{coerces a vector into a}data.frame` object that occupies one row.
}
