% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalglobal-package.R
\docType{package}
\name{personalglobal-package}
\alias{personalglobal}
\alias{personalglobal-package}
\title{personalglobal: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hctchristine/personalglobal}
  \item Report bugs at \url{https://github.com/hctchristine/personalglobal/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
